Refactor: Improve code quality, add docs, and fix critical build issues
