+++ apps/web/src/components/progress/JourneyMap.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useState, useEffect } from 'react';
-import { motion } from 'framer-motion';
+import { motion, AnimatePresence } from 'framer-motion';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
 import { AsciiArt } from '@/components/AsciiArt';
 import { formatDistanceToNow } from 'date-fns';
